Use a single route component
