+++ app/importer/client/routes.js
@@ -2,20 +2,24 @@ import { registerAdminRoute } from '../../ui-admin/client';
 
 registerAdminRoute('/import', {
 	name: 'admin-import',
-	lazyRouteComponent: () => import('./components/ImportHistoryRoute'),
+	lazyRouteComponent: () => import('./components/ImportRoute'),
+	props: { page: 'history' },
 });
 
 registerAdminRoute('/import/new/:importerKey?', {
 	name: 'admin-import-new',
-	lazyRouteComponent: () => import('./components/NewImportRoute'),
+	lazyRouteComponent: () => import('./components/ImportRoute'),
+	props: { page: 'new' },
 });
 
 registerAdminRoute('/import/prepare', {
 	name: 'admin-import-prepare',
-	lazyRouteComponent: () => import('./components/PrepareImportRoute'),
+	lazyRouteComponent: () => import('./components/ImportRoute'),
+	props: { page: 'prepare' },
 });
 
 registerAdminRoute('/import/progress', {
 	name: 'admin-import-progress',
-	lazyRouteComponent: () => import('./components/ImportProgressRoute'),
+	lazyRouteComponent: () => import('./components/ImportRoute'),
+	props: { page: 'progress' },
 });
